refactor(crew): split #615 rescue-placed-marker cluster into CrewReservationManager.RescueMarker.cs
